Deployment/requirements.txt
flask==1.1.2
numpy==1.19.5
pillow==8.1.0
tensorflow=2.4.1